separate renderer and hide details
