+++ vs/superlucent/superlucent.vcxproj
@@ -165,6 +165,7 @@ python compile.py
     <ClCompile Include="..\..\src\main.cc" />
     <ClCompile Include="..\..\src\superlucent\application.cc" />
     <ClCompile Include="..\..\src\superlucent\engine\engine.cc" />
+    <ClCompile Include="..\..\src\superlucent\engine\particle_renderer.cc" />
     <ClCompile Include="..\..\src\superlucent\engine\particle_simulation.cc" />
     <ClCompile Include="..\..\src\superlucent\scene\camera.cc" />
     <ClCompile Include="..\..\src\superlucent\scene\camera_control.cc" />
@@ -173,7 +174,10 @@ python compile.py
   <ItemGroup>
     <ClInclude Include="..\..\src\superlucent\application.h" />
     <ClInclude Include="..\..\src\superlucent\engine\engine.h" />
+    <ClInclude Include="..\..\src\superlucent\engine\particle_renderer.h" />
     <ClInclude Include="..\..\src\superlucent\engine\particle_simulation.h" />
+    <ClInclude Include="..\..\src\superlucent\engine\ubo\camera_ubo.h" />
+    <ClInclude Include="..\..\src\superlucent\engine\ubo\light_ubo.h" />
     <ClInclude Include="..\..\src\superlucent\scene\camera.h" />
     <ClInclude Include="..\..\src\superlucent\scene\camera_control.h" />
     <ClInclude Include="..\..\src\superlucent\scene\light.h" />
